发送前清除串口缓存避免排长队； 修改传给绘图函数的gForce数据类型。

+++ oym8CHWave/mainwindow.h
@@ -43,7 +43,7 @@ private slots:
 
   void on_actionConnect_triggered();
 
-  void on_drawLine(QVector<double> rawData);
+  void on_drawLine(QVector<uint8_t> rawData);
 
 
 protected:
